restructure and split up index to separate files
abide to new lint rules
