+++ .eslintignore
@@ -1,3 +1,3 @@
-dist
-node_modules
-tmp
+dist/
+node_modules/
+tmp/
