+++ .eslintrc.json
@@ -9,28 +9,119 @@
   },
   "extends": "eslint:recommended",
   "rules": {
-    "eol-last": ["error", "always"],
-    "indent": ["error", 2, {"SwitchCase": 1}],
-    "linebreak-style": ["error", "unix"],
-    "no-console": 2,
+    "brace-style": [2, "1tbs", {
+      "allowSingleLine": false
+    }],
+    "camelcase": [
+      "error",
+      {
+        "properties": "never"
+      }
+    ],
+    "comma-dangle": [
+      "warn",
+      {
+        "arrays": "always-multiline",
+        "objects": "always-multiline",
+        "imports": "ignore",
+        "exports": "ignore",
+        "functions": "ignore"
+      }
+    ],
+    "comma-spacing": 2,
+    "consistent-return": 0,
+    "curly": [2, "multi-line"],
+    "dot-notation": [2, {
+      "allowKeywords": true
+    }],
+    "eol-last": 2,
+    "eqeqeq": 2,
+    "handle-callback-err": 2,
+    "indent": [2, 2, {
+      "SwitchCase": 1
+    }],
+    "key-spacing": [2, {
+      "beforeColon": false,
+      "afterColon": true
+    }],
+    "keyword-spacing": 2,
+    "max-len": 0,
+    "new-cap": 0,
+    "new-parens": 2,
+    "no-alert": 2,
+    "no-array-constructor": 2,
+    "no-caller": 2,
+    "no-catch-shadow": 2,
+    "no-console": 1,
+    "no-labels": 2,
+    "no-eval": 2,
+    "no-extend-native": 2,
+    "no-extra-bind": 2,
+    "no-extra-parens": [2, "functions"],
     "no-fallthrough": 0,
+    "no-implied-eval": 2,
+    "no-iterator": 2,
+    "no-label-var": 2,
+    "no-lone-blocks": 2,
+    "no-loop-func": 2,
     "no-multi-spaces": 2,
+    "no-multi-str": 2,
     "no-multiple-empty-lines": [ "error", {
       "max": 1,
       "maxEOF": 0,
       "maxBOF": 0
     } ],
+    "no-native-reassign": 0,
+    "no-new": 0,
+    "no-new-func": 2,
+    "no-new-object": 2,
+    "no-new-wrappers": 2,
+    "no-octal-escape": 2,
+    "no-path-concat": 2,
+    "no-process-exit": 2,
+    "no-proto": 2,
+    "no-prototype-builtins": 2,
+    "no-return-assign": 2,
+    "no-script-url": 2,
+    "no-sequences": 2,
+    "no-shadow": 2,
+    "no-shadow-restricted-names": 2,
+    "no-spaced-func": 2,
+    "no-trailing-spaces": 2,
     "no-undef": 2,
     "no-undef-init": 2,
-    "no-use-before-define": 0,
-    "quotes": ["error", "single"],
-    "require-await": 2,
-    "semi": ["error", "always"],
-    "space-before-function-paren": ["error", {
-      "anonymous": "always",
-      "named": "never",
-      "asyncArrow": "always"
+    "no-underscore-dangle": 0,
+    "no-unused-expressions": 2,
+    "no-use-before-define": ["error", {
+      "functions": false,
+      "classes": true,
+      "variables": true,
+      "allowNamedExports": false
     }],
+    "no-var": 0,
+    "no-with": 2,
+    "prefer-arrow-callback": 0,
+    "prefer-const": ["error", {"destructuring": "all"}],
+    "quotes": [2, "single"],
+    "require-atomic-updates": 0,
+    "require-await": 2,
+    "semi": [2, "always"],
+    "semi-spacing": [2, {
+      "before": false,
+      "after": true
+    }],
+    "space-infix-ops": 2,
+    "space-unary-ops": [2, {
+      "words": true,
+      "nonwords": false
+    }],
+    "space-before-blocks": 2,
+    "space-before-function-paren": [2, {
+      "anonymous": "never",
+      "named": "never"
+    }],
+    "strict": [2, "global"],
+    "yoda": [2, "never"],
     "unicode-bom": ["error", "never"]
   }
 }
