+++ src/Errors.js
@@ -1,4 +1,4 @@
-class NotImplemented extends Error {
+export class NotImplemented extends Error {
   constructor(serviceId) {
     super(`${serviceId} service function not found`);
     this.code = 'EFLOW_NOT_IMPLEMENTED';
@@ -6,6 +6,10 @@ class NotImplemented extends Error {
   }
 }
 
-export {
-  NotImplemented,
-};
+export class FormatError extends Error {
+  constructor(elementId, err) {
+    super(`<${elementId}> ${err.message}`);
+    this.code = 'EFLOW_FORMAT';
+    this.output = {statusCode: 500};
+  }
+}
