+++ src/IO.js
@@ -1,12 +1,115 @@
-export {
-  InputOutput,
-  IOBase,
-  IOMap,
-  IOList,
-  IOScript,
-};
+export class IOBase {
+  constructor(parm) {
+    this.name = parm.name;
+    this.type = parm.definition && parm.definition.$type || 'string';
+    this.behaviour = parm;
+  }
+  getValue(activity, executionMessage) {
+    return {
+      [this.name]: activity.environment.resolveExpression(this.behaviour.value, executionMessage),
+    };
+  }
+}
 
-class InputOutput {
+class IOMap extends IOBase {
+  constructor(parm) {
+    super(parm);
+  }
+  getValue(activity, executionMessage) {
+    const name = this.name;
+    const entries = this.behaviour.definition.entries;
+    if (!Array.isArray(entries)) return {[name]: {}};
+
+    const environment = activity.environment;
+
+    const result = {};
+    for (const {key, value} of entries) {
+      if (!key) continue;
+
+      const val = environment.resolveExpression(value, executionMessage);
+      if (key in result) {
+        if (val === undefined) continue;
+        const current = result[key];
+        if (Array.isArray(current)) {
+          current.push(val);
+        } else {
+          const items = result[key] = [];
+          if (current !== undefined) items.push(current);
+          items.push(val);
+        }
+      } else {
+        result[key] = val;
+      }
+    }
+
+    return {[name]: result};
+  }
+}
+
+export class IOList extends IOBase {
+  constructor(parm) {
+    super(parm);
+  }
+  getValue(activity, executionMessage) {
+    const name = this.name;
+    const items = this.behaviour.definition.items;
+
+    const result = [];
+    if (!Array.isArray(items)) return {[name]: result};
+
+    const environment = activity.environment;
+    for (const item of items) {
+      const val = environment.resolveExpression(item.value, executionMessage);
+      if (val !== undefined) result.push(val);
+    }
+
+    return {
+      [name]: result,
+    };
+  }
+}
+
+export class IOScript extends IOBase {
+  constructor(parm, scriptId) {
+    super(parm);
+    this.id = scriptId;
+    this.script = true;
+  }
+  getValue(activity, executionMessage) {
+    const definition = this.behaviour.definition;
+    const name = this.name;
+    const environment = activity.environment;
+
+    const {fields, content, properties, ...rest} = executionMessage;
+    const scope = {
+      id: this.id,
+      type: this.type,
+      name,
+      fields,
+      content,
+      properties,
+      environment: activity.environment,
+      logger: activity.logger,
+      resolveExpression,
+      ...rest,
+    };
+
+    const script = environment.scripts.getScript(definition.scriptFormat, {id: this.id});
+
+    return new Promise((resolve, reject) => {
+      return script.execute(scope, (err, result) => {
+        if (err) return reject(err);
+        resolve({[name]: result});
+      });
+    });
+
+    function resolveExpression(expression) {
+      return environment.resolveExpression(expression, scope);
+    }
+  }
+}
+
+export class InputOutput {
   constructor(parentId, behaviour, context) {
     this.parentId = parentId;
     this.context = context;
@@ -66,116 +169,5 @@ class InputOutput {
     }
 
     return mapped;
-  }
-}
-
-class IOBase {
-  constructor(parm) {
-    this.name = parm.name;
-    this.type = parm.definition && parm.definition.$type || 'string';
-    this.behaviour = parm;
-  }
-  getValue(activity, executionMessage) {
-    return {
-      [this.name]: activity.environment.resolveExpression(this.behaviour.value, executionMessage),
-    };
-  }
-}
-
-class IOMap extends IOBase {
-  constructor(parm) {
-    super(parm);
-  }
-  getValue(activity, executionMessage) {
-    const name = this.name;
-    const entries = this.behaviour.definition.entries;
-    if (!Array.isArray(entries)) return {[name]: {}};
-
-    const environment = activity.environment;
-
-    const result = {};
-    for (const {key, value} of entries) {
-      if (!key) continue;
-
-      const val = environment.resolveExpression(value, executionMessage);
-      if (key in result) {
-        if (val === undefined) continue;
-        const current = result[key];
-        if (Array.isArray(current)) {
-          current.push(val);
-        } else {
-          const items = result[key] = [];
-          if (current !== undefined) items.push(current);
-          items.push(val);
-        }
-      } else {
-        result[key] = val;
-      }
-    }
-
-    return {[name]: result};
-  }
-}
-
-class IOList extends IOBase {
-  constructor(parm) {
-    super(parm);
-  }
-  getValue(activity, executionMessage) {
-    const name = this.name;
-    const items = this.behaviour.definition.items;
-
-    const result = [];
-    if (!Array.isArray(items)) return {[name]: result};
-
-    const environment = activity.environment;
-    for (const item of items) {
-      const val = environment.resolveExpression(item.value, executionMessage);
-      if (val !== undefined) result.push(val);
-    }
-
-    return {
-      [name]: result,
-    };
-  }
-}
-
-class IOScript extends IOBase {
-  constructor(parm, scriptId) {
-    super(parm);
-    this.id = scriptId;
-    this.script = true;
-  }
-  getValue(activity, executionMessage) {
-    const definition = this.behaviour.definition;
-    const name = this.name;
-    const environment = activity.environment;
-
-    const {fields, content, properties, ...rest} = executionMessage;
-    const scope = {
-      id: this.id,
-      type: this.type,
-      name,
-      fields,
-      content,
-      properties,
-      environment: activity.environment,
-      logger: activity.logger,
-      resolveExpression,
-      ...rest,
-    };
-
-    const script = environment.scripts.getScript(definition.scriptFormat, {id: this.id});
-
-    return new Promise((resolve, reject) => {
-      return script.execute(scope, (err, result) => {
-        if (err) return reject(err);
-        resolve({[name]: result});
-      });
-    });
-
-    function resolveExpression(expression) {
-      return environment.resolveExpression(expression, scope);
-    }
   }
 }
