+++ test/src/ServiceExpression-test.js
@@ -5,8 +5,8 @@ describe('ServiceExpression', () => {
     ServiceExpression({
       type: 'bpmn:ServiceTask',
       behaviour: {
-        expression: '${expr}'
-      }
+        expression: '${expr}',
+      },
     });
   });
 });
